Use int instead of u8 as config parameter type
Functions for handling the int configuration parameters use int pointers
and a u8 value here could result in unaligned operation.

+++ wpa_supplicant/config_ssid.h
@@ -1380,7 +1380,7 @@ struct wpa_ssid {
 	 * 1 = check AP's extended capabilities
 	 * 2 = always enable
 	 */
-	u8 always_use_proxy_arp;
+	int always_use_proxy_arp;
 };
 
 #endif /* CONFIG_SSID_H */
